rename: BookSummaryDto 패키지 변경 (book → bookshelf) (#14)
