+++ bookiary-api/src/main/java/oz/bookiarybacked/domain/bookshelf/domain/repository/UserBookQueryRepository.java
@@ -2,7 +2,8 @@ package oz.bookiarybacked.domain.bookshelf.domain.repository;
 
 import oz.bookiarybacked.common.dto.Page;
 import oz.bookiarybacked.common.dto.PageParam;
-import oz.bookiarybacked.domain.book.domain.dto.BookSummaryDto;
+import oz.bookiarybacked.domain.bookshelf.domain.dto.BookDetailDto;
+import oz.bookiarybacked.domain.bookshelf.domain.dto.BookSummaryDto;
 
 public interface UserBookQueryRepository {
 	Page<BookSummaryDto> retrieveBookshelf(Long userId, PageParam pageParam);
