+++ bookiary-api/src/main/java/oz/bookiarybacked/infra/query/QuerydslUserBookQueryRepository.java
@@ -15,7 +15,8 @@ import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import oz.bookiarybacked.common.dto.Page;
 import oz.bookiarybacked.common.dto.PageParam;
-import oz.bookiarybacked.domain.book.domain.dto.BookSummaryDto;
+import oz.bookiarybacked.domain.bookshelf.domain.dto.BookDetailDto;
+import oz.bookiarybacked.domain.bookshelf.domain.dto.BookSummaryDto;
 import oz.bookiarybacked.domain.bookshelf.domain.repository.UserBookQueryRepository;
 
 @Repository
